+++ polardb-20170801/alibabacloud_polardb20170801/models/_describe_user_encryption_key_list_response_body.py
@@ -11,14 +11,22 @@ class DescribeUserEncryptionKeyListResponseBody(DaraModel):
         self,
         dbcluster_id: str = None,
         key_list: List[str] = None,
+        page_number: int = None,
+        page_record_count: int = None,
+        page_size: int = None,
         request_id: str = None,
+        total_record_count: int = None,
     ):
         # The ID of the cluster.
         self.dbcluster_id = dbcluster_id
         # Cluster key list.
         self.key_list = key_list
+        self.page_number = page_number
+        self.page_record_count = page_record_count
+        self.page_size = page_size
         # The ID of the request.
         self.request_id = request_id
+        self.total_record_count = total_record_count
 
     def validate(self):
         pass
@@ -34,8 +42,20 @@ class DescribeUserEncryptionKeyListResponseBody(DaraModel):
         if self.key_list is not None:
             result['KeyList'] = self.key_list
 
+        if self.page_number is not None:
+            result['PageNumber'] = self.page_number
+
+        if self.page_record_count is not None:
+            result['PageRecordCount'] = self.page_record_count
+
+        if self.page_size is not None:
+            result['PageSize'] = self.page_size
+
         if self.request_id is not None:
             result['RequestId'] = self.request_id
+
+        if self.total_record_count is not None:
+            result['TotalRecordCount'] = self.total_record_count
 
         return result
 
@@ -47,8 +67,20 @@ class DescribeUserEncryptionKeyListResponseBody(DaraModel):
         if m.get('KeyList') is not None:
             self.key_list = m.get('KeyList')
 
+        if m.get('PageNumber') is not None:
+            self.page_number = m.get('PageNumber')
+
+        if m.get('PageRecordCount') is not None:
+            self.page_record_count = m.get('PageRecordCount')
+
+        if m.get('PageSize') is not None:
+            self.page_size = m.get('PageSize')
+
         if m.get('RequestId') is not None:
             self.request_id = m.get('RequestId')
+
+        if m.get('TotalRecordCount') is not None:
+            self.total_record_count = m.get('TotalRecordCount')
 
         return self